Preview unit in queue while materializing
Unit's avatar will show in queue while player chooses the location
for materializing it. The avatar will have 50% opacity until the
location has been confirmed and it will become fully opaque.

The avatar will disappear from the queue when the player skips
their turn, when the player chooses a different unit to
materialize, or when the player cancels the materialize ability.